[REFACTORE]: fixed locale emulator and removed duplicated files
