+++ Codes/backend/functions/.gitignore
@@ -41,3 +41,4 @@ testem.log
 # System files
 .DS_Store
 Thumbs.db
+*.log
