+++ Codes/backend/functions/src/index.ts
@@ -4,7 +4,8 @@ import { sendPasswordReset } from './callables/auth/sendPasswordReset';
 import { registerUser } from './callables/auth/registerUser';
 import { changeUserEmail } from './callables/auth/changeUserEmail';
 import { checkServiceStatus } from './callables/auth/checkServiceStatus';
-
+import { login } from './callables/auth/login';
+import { logout } from './callables/auth/logout';
 // Export callables
 export {
   sendPasswordReset,
@@ -12,4 +13,6 @@ export {
   registerUser,
   changeUserEmail,
   checkServiceStatus,
+  login,
+  logout,
 };
